blog post displayed

+++ src/utils/helperFunctions.js
@@ -1,20 +1,31 @@
 export const setToLocalStorage = (key, value) => {
-  localStorage.setItem(key, JSON.stringify(value));
+  if (typeof window !== "undefined") {
+    localStorage.setItem(key, JSON.stringify(value));
+  }
 };
 
 export const getFromLocalStorage = (key) => {
-  const itemString = localStorage.getItem(key);
-  return itemString ? JSON.parse(itemString) : null;
+  if (typeof window !== "undefined") {
+    const itemString = localStorage.getItem(key);
+    return itemString ? JSON.parse(itemString) : null;
+  }
 };
 
 export const removeFromLocalStorage = (key) => {
-  localStorage.removeItem(key);
+  if (typeof window !== "undefined") {
+    localStorage.removeItem(key);
+  }
 };
 
 export const isItemInLocalStorage = (key) => {
-  return localStorage.getItem(key) !== null;
+  if (typeof window !== "undefined") {
+    return localStorage.getItem(key) !== null;
+  }
+  return false;
 };
 
 export const clearLocalStorage = () => {
-  localStorage.clear();
+  if (typeof window !== "undefined") {
+    localStorage.clear();
+  }
 };
